refactor(Core): Changement des connexions de base de données - Migration des tables vers leurs bases de données respectives
- Modification des fichiers de migration : Création des tables dans leurs bases de données respectives
- Mise à jour des modèles : Ajout de la propriété de connexion appropriée à chaque modèle
- Suppression du fichier de migration de la table 'authentication_log'

+++ database/migrations/2024_01_29_205300_create_acquaintances_friendship_table.php
@@ -8,19 +8,20 @@ class CreateAcquaintancesFriendshipTable extends Migration
 {
     public function up()
     {
-
-        Schema::create(config('acquaintances.tables.friendships'), function (Blueprint $table) {
-            $table->id();
-            $table->morphs('sender');
-            $table->morphs('recipient');
-            $table->string('status')->default('pending')->comment('pending/accepted/denied/blocked/');
-            $table->timestamps();
-        });
+        if (DB::connection()->getDriverName() == 'mysql')
+            Schema::create(config('acquaintances.tables.friendships'), function (Blueprint $table) {
+                $table->id();
+                $table->morphs('sender');
+                $table->morphs('recipient');
+                $table->string('status')->default('pending')->comment('pending/accepted/denied/blocked/');
+                $table->timestamps();
+            });
 
     }
 
     public function down()
     {
-        Schema::dropIfExists(config('acquaintances.tables.friendships'));
+        if (DB::connection()->getDriverName() == 'mysql')
+            Schema::dropIfExists(config('acquaintances.tables.friendships'));
     }
 }
